+++ src/components/ChatWindow.tsx
@@ -1,5 +1,6 @@
-import {useState} from 'react'
+import React, {useState} from 'react'
 import WebSocketWrapper from './WebSocketWrapper';
+import {Paper} from '@mui/material'
 import ForumIcon from '@mui/icons-material/Forum';
 
 const ChatWindow = () =>{
@@ -8,9 +9,11 @@ const ChatWindow = () =>{
 
     return (
         <div className="chatWindowContainer">
-            {!openChat ? <ForumIcon onClick={()=>setOpenChat(true)} fontSize="large">Start Chatting!</ForumIcon> : null }
+            {!openChat ? <ForumIcon data-testid="ForumIcon" onClick={()=>setOpenChat(true)} fontSize="large">Start Chatting!</ForumIcon> : null }
             {openChat
-                ?   <WebSocketWrapper closeChat={()=>setOpenChat(false)} />
+                ?   <Paper elevation={24} style={{padding: 20}}>
+                        <WebSocketWrapper closeChat={()=>setOpenChat(false)} />
+                    </Paper>
                 :   null
         }
         </div>